Added the first working version of the game
The Tic-Tac_Toe game for two players, X and O, who take turns marking
the spaces in a 3×3 grid. The player who succeeds in placing three of
their marks in a horizontal, vertical, or diagonal row is the winner.

After starting the game, an empty playing field and a line appear in
front of you, inviting you to enter coordinates. Player H moves first,
he must select an empty cell on the playing field by entering
coordinates in the x y format, where each coordinate must be in the
range from 1 to 3.

The layout of the field starts from the lower left corner,
which has coordinates 1 1, lower right the corner has coordinates 3 1.

Let's play, good luck to the players.